Move app-registry service to service tracer.
This changes the app-registry service away from using global tracers,
which makes debugging tracing issues easier going forward.
